+++ Home.py
@@ -1,94 +1,23 @@
 import os
 from threading import Lock
-import pandas as pd
+
 import streamlit as st
-from requests import ConnectionError, HTTPError
-import time
+
 from globals import VARS_INDEX_PATH
+from widgets.console import show_console
 from widgets.index import (
     get_last_index_update,
-    load_table_of_contents,
     load_codelist_reverse_index,
+    load_table_of_contents,
+    save_index_file,
 )
-from src.eurostat import (
-    eurostat_sdmx_request,
-    fetch_dataset_codelist,
-)
-from widgets.console import show_console
 from widgets.session import app_config
-
-
-@st.experimental_singleton(show_spinner=False)
-def api_endpoint():
-    return eurostat_sdmx_request()
 
 
 @st.experimental_singleton(show_spinner=False)
 def index_lock():
     """A shared lock amongst sessions to prevent concurrent index write."""
     return Lock()
-
-
-def save_index_file():
-    """Obtain codelist and in which dataset is used each."""
-    message = st.sidebar.empty()
-    message.text("Initializing indexing...")
-
-    codelist = pd.DataFrame()
-    progress_bar = st.sidebar.progress(0.0)
-    progress_value = 0.0
-    status = st.sidebar.empty()
-    req = api_endpoint()
-    datasets = load_table_of_contents().index.to_list()
-    len_datasets = len(datasets)
-    datasets_not_loaded = []
-    for n, dataset in enumerate(datasets):
-        message.text(f"Loading {dataset} ({n}/{len_datasets})")
-        try:
-            codes, cached = fetch_dataset_codelist(req, dataset)
-            codes = codes.assign(dataset=dataset)
-            codelist = pd.concat([codelist, codes])
-            status.success(
-                f"Loaded{' from cache' if cached else ' from Eurostat'}",
-                icon="♻️" if cached else "✅",
-            )
-        except HTTPError as e:
-            # NOTE Usually happens if a dataset was not found
-            status.warning(e, icon="⚠️")
-            datasets_not_loaded.append(dataset)
-        except ConnectionError as e:
-            # NOTE Usually happens when Eurostat reset connection
-            status.warning(e, icon="⚠️")
-            datasets_not_loaded.append(dataset)
-            time.sleep(1)  # Cooldown before resume requests
-        progress_value = n / len_datasets
-        progress_bar.progress(progress_value)
-
-    progress_bar.empty()
-    if len(datasets_not_loaded) > 0:
-        status.warning(
-            f"Unable to load: {datasets_not_loaded}. Refresh only if you find it useful. All the metadata successfully received will be loaded from cache.",
-            icon="⚠️",
-        )
-    else:
-        status.success(
-            f"All datasets were loaded.",
-            icon="✅",
-        )
-
-    # Aggregate datasets dimension for a reverse index
-    message.text("Finalizing indexing...")
-    codelist = (
-        codelist.assign(name=codelist.name.str.capitalize())
-        .groupby(["dimension", "name"])["dataset"]
-        .unique()
-    )
-    codelist = codelist[codelist.apply(len) > 0]
-    codelist.index = codelist.index.to_flat_index().str.join(" | ")
-    codelist.name = "datasets"
-    codelist.index.name = "code"
-    codelist.apply(lambda x: x.tolist()).to_pickle(VARS_INDEX_PATH)
-    message.empty()
 
 
 def index_helper(message_widget):
