+++ src/eurostat.py
@@ -21,8 +21,8 @@ def fetch_table_of_contents() -> pd.Series:
     """Returns dataset codes as keys and titles as values."""
     # NOTE Access to txt seems quicker than a SDMX call:
     # `dataflows = pandasdmx.to_pandas(eurostat_sdmx_request().dataflow())`
-    # It's also returning more results than in the txt!
-    return (
+    # It's also returning more results than in the txt and that can't be found in bulk download.
+    toc = (
         pd.read_table(
             "https://ec.europa.eu/eurostat/estat-navtree-portlet-prod/BulkDownloadListing?file=table_of_contents_en.txt",
             usecols=[0, 1, 2],
@@ -31,11 +31,12 @@ def fetch_table_of_contents() -> pd.Series:
         .query("type == 'dataset'")
         .drop(columns=["type"])
         .transform(lambda x: x.str.strip())
+        .drop_duplicates()
         .set_index("code")
         .squeeze()
         .sort_index()
-        .drop_duplicates()
     )
+    return toc
 
 
 def fetch_dataset_codelist(
